feat : 파일구조 적용 및 일부 entity 코드 수정
