+++ src/main/java/org/cobee/server/alarm/Enum/AlarmType.java
@@ -1,4 +1,4 @@
-package org.cobee.server.Enum;
+package org.cobee.server.alarm.Enum;
 
 public enum AlarmType {
     COMMENT,
