+++ src/main/java/org/cobee/server/alarm/controller/AlarmController.java
@@ -1,0 +1,9 @@
+package org.cobee.server.alarm.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class AlarmController {
+}
