+++ src/main/java/org/cobee/server/alarm/entity/Alarm.java
@@ -1,30 +1,32 @@
-package org.cobee.server.alarm;
+package org.cobee.server.alarm.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.cobee.server.alarm.Enum.AlarmType;
 import org.cobee.server.member.Member;
 
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
-public class AlarmNotice {
+public class Alarm {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column
-    private Boolean is_read;
+    @Enumerated(EnumType.STRING)
+    private AlarmType alarmType;
+
+    @Column
+    private int where_from; // 왜 int??
 
     @ManyToOne
-    @JoinColumn(name="to_user_id")
+    @JoinColumn(name="from_user_id")
     private Member member;
 
-    @ManyToOne
-    @JoinColumn(name="alarm_id")
-    private Alarm alarm;
 
 }
