+++ src/main/java/org/cobee/server/alarm/entity/AlarmNotice.java
@@ -1,35 +1,31 @@
-package org.cobee.server.comment;
+package org.cobee.server.alarm.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.cobee.server.alarm.entity.Alarm;
 import org.cobee.server.member.Member;
-import org.cobee.server.recruit.RecruitPost;
 
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
-public class Comment {
+public class AlarmNotice {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(columnDefinition = "TEXT")
-    private String content;
-
     @Column
-    private Boolean is_private;
+    private Boolean is_read;
 
     @ManyToOne
-    @JoinColumn(name="user_id")
+    @JoinColumn(name="to_user_id")
     private Member member;
 
     @ManyToOne
-    @JoinColumn(name = "post_id")
-    private RecruitPost post;
-
+    @JoinColumn(name="alarm_id")
+    private Alarm alarm;
 
 }
