+++ src/main/java/org/cobee/server/alarm/repository/AlarmNoticeRepository.java
@@ -1,0 +1,4 @@
+package org.cobee.server.alarm.repository;
+
+public interface AlarmNoticeRepository {
+}
