+++ src/main/java/org/cobee/server/alarm/repository/AlarmRepository.java
@@ -1,0 +1,4 @@
+package org.cobee.server.alarm.repository;
+
+public interface AlarmRepository {
+}
