+++ src/main/java/org/cobee/server/alarm/service/AlarmService.java
@@ -1,0 +1,7 @@
+package org.cobee.server.alarm.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class AlarmService {
+}
