+++ src/main/java/org/cobee/server/chat/controller/ChattingRoomController.java
@@ -1,0 +1,9 @@
+package org.cobee.server.chat.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class ChattingRoomController {
+}
