+++ src/main/java/org/cobee/server/chat/repository/ChattingRoomRepository.java
@@ -1,0 +1,4 @@
+package org.cobee.server.chat.repository;
+
+public interface ChattingRoomRepository {
+}
