+++ src/main/java/org/cobee/server/chat/service/ChattingRoomService.java
@@ -1,0 +1,4 @@
+package org.cobee.server.chat.service;
+
+public class ChattingRoomService {
+}
