+++ src/main/java/org/cobee/server/comment/controller/CommentController.java
@@ -1,0 +1,9 @@
+package org.cobee.server.comment.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class CommentController {
+}
