+++ src/main/java/org/cobee/server/comment/repository/CommentRepository.java
@@ -1,0 +1,4 @@
+package org.cobee.server.comment.repository;
+
+public interface CommentRepository {
+}
