+++ src/main/java/org/cobee/server/comment/service/CommentService.java
@@ -1,0 +1,7 @@
+package org.cobee.server.comment.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class CommentService {
+}
