+++ src/main/java/org/cobee/server/common/Enum/Gender.java
@@ -1,4 +1,4 @@
-package org.cobee.server.Enum;
+package org.cobee.server.common.Enum;
 
 public enum Gender {
     MALE, FEMALE
