+++ src/main/java/org/cobee/server/common/Enum/Lifecycle.java
@@ -1,4 +1,4 @@
-package org.cobee.server.Enum;
+package org.cobee.server.common.Enum;
 
 public enum Lifecycle {
     MORNING, EVENING
