+++ src/main/java/org/cobee/server/common/Enum/RecruitStatus.java
@@ -1,4 +1,4 @@
-package org.cobee.server.Enum;
+package org.cobee.server.common.Enum;
 
 public enum RecruitStatus {
     // 모집중, 컨택중, 모집완료
