+++ src/main/java/org/cobee/server/common/Enum/SocialType.java
@@ -1,4 +1,4 @@
-package org.cobee.server.Enum;
+package org.cobee.server.common.Enum;
 
 public enum SocialType {
     KAKAO, GOOGLE
