+++ src/main/java/org/cobee/server/config/CorsConfig.java
@@ -1,0 +1,4 @@
+package org.cobee.server.config;
+
+public class CorsConfig {
+}
