+++ src/main/java/org/cobee/server/config/SwaggerConfig.java
@@ -1,0 +1,4 @@
+package org.cobee.server.config;
+
+public class SwaggerConfig {
+}
