+++ src/main/java/org/cobee/server/global/response/ApiResponse.java
@@ -1,0 +1,4 @@
+package org.cobee.server.global.response;
+
+public class ApiResponse {
+}
