+++ src/main/java/org/cobee/server/recruit/controller/RecruitController.java
@@ -1,0 +1,9 @@
+package org.cobee.server.recruit.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class RecruitController {
+}
