+++ src/main/java/org/cobee/server/recruit/repository/RecruitPostRepository.java
@@ -1,0 +1,9 @@
+package org.cobee.server.recruit.repository;
+
+import org.cobee.server.recruit.entity.RecruitPost;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RecruitPostRepository extends JpaRepository<RecruitPost, Long> {
+}
