+++ src/main/java/org/cobee/server/recruit/service/RecruitService.java
@@ -1,0 +1,9 @@
+package org.cobee.server.recruit.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class RecruitService {
+}
